Move React hooks out of component modules
