+++ packages/react/src/ConfirmationDialog/ConfirmationDialog.test.tsx
@@ -6,7 +6,8 @@ import {useCallback, useRef, useState} from 'react'
 import {ActionMenu} from '../deprecated/ActionMenu'
 import BaseStyles from '../BaseStyles'
 import {Button} from '../Button'
-import {ConfirmationDialog, useConfirm} from './ConfirmationDialog'
+import {ConfirmationDialog} from './ConfirmationDialog'
+import {useConfirm} from './useConfirm'
 import {Stack} from '../Stack'
 import {implementsClassName} from '../utils/testing'
 import dialogClasses from '../Dialog/Dialog.module.css'
