+++ packages/react/src/experimental/Tabs/Tabs.examples.stories.tsx
@@ -1,7 +1,9 @@
 import type {Meta} from '@storybook/react-vite'
 import {action} from 'storybook/actions'
 import React from 'react'
-import {Tabs, TabPanel, useTabList, useTab} from './Tabs'
+import {Tabs, TabPanel} from './Tabs'
+import {useTab} from './useTab'
+import {useTabList} from './useTabList'
 import {ActionList} from '../../ActionList'
 import Flash from '../../Flash'
