+++ packages/react/src/experimental/Tabs/TabsContext.ts
@@ -1,0 +1,4 @@
+import {createContext} from 'react'
+import type {TabsContextValue} from './types'
+
+export const TabsContext = createContext<TabsContextValue | null>(null)
