+++ packages/react/src/index.ts
@@ -3,7 +3,8 @@
 export {default as theme, type ThemeColorPaths, type ThemeShadowPaths} from './theme'
 export {default as BaseStyles} from './BaseStyles'
 export type {BaseStylesProps} from './BaseStyles'
-export {default as ThemeProvider, useTheme, useColorSchemeVar} from './ThemeProvider'
+export {default as ThemeProvider} from './ThemeProvider'
+export {useTheme, useColorSchemeVar} from './useTheme'
 export type {ThemeProviderProps} from './ThemeProvider'
 
 // Layout
@@ -33,7 +34,7 @@ export type {TouchOrMouseEvent} from './hooks/useOnOutsideClick'
 export {useOpenAndCloseFocus} from './hooks/useOpenAndCloseFocus'
 export {useOnEscapePress} from './hooks/useOnEscapePress'
 export {useOverlay} from './hooks/useOverlay'
-export {useConfirm} from './ConfirmationDialog/ConfirmationDialog'
+export {useConfirm} from './ConfirmationDialog/useConfirm'
 export {useFocusTrap} from './hooks/useFocusTrap'
 export type {FocusTrapHookSettings} from './hooks/useFocusTrap'
 export {FocusKeys, useFocusZone} from './hooks/useFocusZone'
