+++ packages/react/src/stories/ThemeProvider.stories.tsx
@@ -1,6 +1,7 @@
 import type {Meta, StoryFn} from '@storybook/react-vite'
 
-import {ThemeProvider, type ThemeProviderProps, useTheme} from '../ThemeProvider'
+import ThemeProvider, {type ThemeProviderProps} from '../ThemeProvider'
+import {useTheme} from '../useTheme'
 import BaseStyles from '../BaseStyles'
 import classes from './ThemeProvider.stories.module.css'
